+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Changelog structure reference:
 #### Zon
 - Basic parser and PSI tree
 - Basic syntax highlighting
+- Color settings page
 
 ## [0.4.0]